utility_scripts/requirements.txt
langchain
openai
torch
transformers
sentence-transformers
pdfplumber
psycopg2-binary
python-docx
